+++ paddle2onnx/op_mapper/detection/yolo_box.py
@@ -17,7 +17,7 @@ from __future__ import absolute_import
 import numpy as np
 from paddle2onnx.constant import dtypes
 from paddle2onnx.op_mapper import OpMapper as op_mapper
-from paddle2onnx import utils
+from paddle2onnx.op_mapper import mapper_helper
 
 MAX_FLOAT32 = np.asarray(
     [255, 255, 127, 127], dtype=np.uint8).view(np.float32)[0]
@@ -38,7 +38,7 @@ class YOLOBox():
     def front(cls, graph, node, **kw):
         model_name = node.output('Boxes', 0)
         input_shape = node.input_shape('X', 0)
-        utils.is_static_shape(input_shape)
+        mapper_helper.is_static_shape(input_shape)
         image_size = node.input('ImgSize')
         input_height = input_shape[2]
         input_width = input_shape[3]
